backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
redis==5.0.1
pydantic==2.6.0
google-generativeai==0.3.2
spacy==3.7.0
networkx==3.2.0
numpy==1.26.0
scipy==1.12.0
scikit-learn==1.4.0
xgboost==2.0.3
simpy==4.1.1
pulp==2.7.0
httpx==0.27.0
beautifulsoup4==4.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.0
neo4j==5.17.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
